Se puede comprar por KG,MONTO Y UNIDAD

+++ Panaderia.Application/DTOs/Venta/ProductoMasVendidoDTO.cs
@@ -9,7 +9,7 @@ namespace Panaderia.Application.DTOs.Ventas
     public class ProductoMasVendidoDto
     {
         public string Nombre { get; set; }
-        public int CantidadVendida { get; set; }
+        public decimal CantidadVendida { get; set; }
         public decimal TotalFacturado { get; set; }
     }
 }
