+++ Panaderia.Domain/Entidades/MovimientoStock.cs
@@ -14,7 +14,7 @@ namespace Panaderia.Domain.Entidades
         public DateTime? Fecha { get; set; } = DateTime.UtcNow;
         public int ProductoId { get; set; }
         public TipoMovimiento Tipo { get; set; }
-        public int Cantidad { get; set; } // Siempre registrar en valor absoluto (positivo)
+        public decimal Cantidad { get; set; } // Siempre registrar en valor absoluto (positivo)
         public string? Observaciones { get; set; }
         public int? ReferenciaId { get; set; } // ID de la Compra o Venta que originó el movimiento
