+++ Panaderia.Domain/Entidades/Ventas/DetalleVenta.cs
@@ -19,9 +19,10 @@ namespace Panaderia.Domain.Entidades.Ventas
 
         public Producto Producto { get; set; }
 
-        public int Cantidad { get; set; }
+        public decimal Cantidad { get; set; }
 
         public decimal PrecioUnitario { get; set; }
+
         public decimal Subtotal { get; set; }
     }
 }
